Move `TypedException` to `core.injector` package and fix imports
